Added SPICE auth
This allows a larger pool of accounts to be used. However, this is limited to only students for now since we *DO NOT* have access to a lecturer account.
A backdoor account(heh) is built in for debugging purposes. You may ask, "Why a backdoor account? What are you doing Daniel?" Simple reasoning.
This allows people who are not registered as a Singapore Poly student to test this software. It also allows others to use it outside of SP.
(Yes I know I can scrape https://esp.sp.edu.sg and use that for auth. but since Justin has already made a wrapper for ATS_NativeUI project, why not borrow the code over there?)